Create and use block-level form-for component
