+++ app/models/sign-in-attempt.js
@@ -1,6 +1,6 @@
 import DS from 'ember-data';
 
 export default DS.Model.extend({
-  email: DS.attr('string'),
-  password: DS.attr('string')
+  email: DS.attr('string', { defaultValue: '' }),
+  password: DS.attr('string', { defaultValue: '' })
 });
