+++ app/templates/components/form-for.hbs
@@ -1,0 +1,6 @@
+<form {{action "validateAndSubmit" on="submit"}}>
+  {{yield (hash
+    changeset=changeset
+    hasAttemptedSubmit=hasAttemptedSubmit
+  )}}
+</form>
