--- app/templates/components/sign-in-form.hbs
@@ -1,28 +1,0 @@
-<form {{action "validateAndSubmit" on="submit"}}>
-  <div class="form-group">
-    {{validated-input
-      autocomplete="email"
-      autofocus=true
-      errors=changeset.errors
-      hasAttemptedSubmit=hasAttemptedSubmit
-      label="Email"
-      name="email"
-      placeholder="user@domain.com"
-      type="email"
-      value=changeset.email
-    }}
-  </div>
-  <div class="form-group">
-    {{validated-input
-      autocomplete="current-password"
-      errors=changeset.errors
-      hasAttemptedSubmit=hasAttemptedSubmit
-      label="Password"
-      name="password"
-      type="password"
-      value=changeset.password
-    }}
-  </div>
-
-  <input data-test-sign-in-submit type="submit" class="btn btn-primary" value="Sign In" />
-</form>
